src/redis_vector_db/requirements.txt
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
python-dotenv==1.0.1
redis==5.0.3
redisvl==0.1.3
requests==2.31.0
sentence-transformers==2.6.1
transformers==4.39.3
